Update methode of handling arguments
